Move model accessor to top level component package
