+++ src/main/common/components/model-accessor.ts
