+++ src/main/common/components/payment-intention/payment-date.ts
@@ -1,7 +1,7 @@
 import * as express from 'express'
 import { Moment } from 'moment'
 
-import { AbstractModelAccessor } from 'shared/components/payment-intention/model-accessor'
+import { AbstractModelAccessor } from 'shared/components/model-accessor'
 import { PaymentIntention } from 'shared/components/payment-intention/model'
 import { Paths as PaymentIntentionPaths } from 'shared/components/payment-intention/paths'
