+++ src/main/features/claimant-response/routes/payment-plan.ts
@@ -1,7 +1,7 @@
 import * as express from 'express'
 
 import { AbstractPaymentPlanPage } from 'shared/components/payment-intention/payment-plan'
-import { AbstractModelAccessor, DefaultModelAccessor } from 'shared/components/payment-intention/model-accessor'
+import { AbstractModelAccessor, DefaultModelAccessor } from 'shared/components/model-accessor'
 import { PaymentIntention } from 'shared/components/payment-intention/model'
 
 import { DraftClaimantResponse } from 'claimant-response/draft/draftClaimantResponse'
