+++ src/main/features/response/routes/full-admission/payment-plan.ts
@@ -1,7 +1,7 @@
 import * as express from 'express'
 
 import { AbstractPaymentPlanPage } from 'shared/components/payment-intention/payment-plan'
-import { AbstractModelAccessor } from 'shared/components/payment-intention/model-accessor'
+import { AbstractModelAccessor } from 'shared/components/model-accessor'
 import { PaymentIntention } from 'shared/components/payment-intention/model'
 import { OptInFeatureToggleGuard } from 'guards/optInFeatureToggleGuard'
