+++ src/main/features/response/routes/partial-admission/payment-date.ts
@@ -1,7 +1,7 @@
 import * as express from 'express'
 
 import { AbstractPaymentDatePage } from 'shared/components/payment-intention/payment-date'
-import { AbstractModelAccessor } from 'shared/components/payment-intention/model-accessor'
+import { AbstractModelAccessor } from 'shared/components/model-accessor'
 import { PaymentIntention } from 'shared/components/payment-intention/model'
 import { OptInFeatureToggleGuard } from 'guards/optInFeatureToggleGuard'
